feat: move resume types within resume data
